Updated demo

+++ PagedListDemo/Models/BooksModel/BooksModel.cs
@@ -6,14 +6,14 @@ namespace PagedListDemo.Models.BooksModel
     public class BooksModel
     {
         [JsonProperty(PropertyName = "id")]
-        public int Id { get; set; }
-
+        public long Id { get; set; }
+        
         [JsonProperty(PropertyName = "title")]
         public string Title { get; set; }
-
+        
         [JsonProperty(PropertyName = "description")]
         public string Description { get; set; }
-
+        
         [JsonProperty(PropertyName = "author")]
         public string Author { get; set; }
